backend/requirements.txt
django
djangorestframework
djoser==2.1.0
drf-yasg==1.21.5
django-cors-headers
Pillow==9.3.0
django-filter==21.1